Update the analytic engine services
