+++ analytic-engine/src/main/java/com/mycompany/jobs/SchemaInferenceJob.java
@@ -3,18 +3,16 @@ package com.mycompany.jobs;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.model.S3Object;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Inject;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import com.mycompany.models.*;
 import com.mycompany.services.bi.BiRepository;
-import com.mycompany.services.mongodb.MongodbRepository;
+import com.mycompany.services.MongodbRepository;
 import org.apache.spark.sql.*;
 import org.apache.spark.sql.types.StructField;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.json.simple.JSONObject;
-import org.mortbay.util.ajax.JSON;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
