+++ analytic-engine/src/main/java/com/mycompany/services/HttpService.java
@@ -1,28 +1,20 @@
 package com.mycompany.services;
 
-import com.google.inject.Inject;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
 import com.mycompany.configuration.Log;
 import com.mycompany.models.ConfigModel;
 
-import java.util.Base64;
 import java.util.Map;
 
-/**
- * GET is used widely in the program.
- * POST and DELETE are only used by the EndToEnd test to
- * set up the environment to run the engine.
- */
 public class HttpService implements Log {
-    private String jobId;
-    private final ConfigModel configModel;
+    private ConfigModel configModel;
+    private String jodId;
 
-    @Inject
     public HttpService(ConfigModel configModel, String jobId) {
         this.configModel = configModel;
-        this.jobId = jobId;
+        this.jodId = jobId;
     }
 
     public HttpResponse<String> get(String url, Map<String, String> headers) throws UnirestException {
